bb generator and EDA_copy
Bounding box generator uses code found on GitHub to create bounding boxes on the dataset

EDA_preprocessing_copy copies the code Geoffrey provided, but it is modified for the size changes required for the ViT model. Will be modifying it again to create image size for YOLO model as well.